wrapping state: deleted course
