+++ src/App.tsx
@@ -8,8 +8,8 @@ import { Welcome } from "./WelcomeMsg";
 import { Button } from "react-bootstrap";
 import { AddPlanModal } from "./components/AddPlanModal";
 import { Drag } from "./components/Drag";
-import { Semester } from "./interfaces/semester";
-import { Course } from "./interfaces/course";
+//import { Semester } from "./interfaces/semester";
+//import { Course } from "./interfaces/course";
 
 //extract plans from data
 const PLANS = semesterPlan.map(
@@ -62,51 +62,6 @@ function App(): JSX.Element {
     const handleCloseAddModal = () => setShowAddModal(false);
     const handleShowAddModal = () => setShowAddModal(true);
 
-    function updateDeletedCourse(
-        planId: string,
-        semId: string,
-        courseCode: string
-    ) {
-        const currPlan = plans.find(
-            (plan: Plan): boolean => plan.id === planId
-        );
-
-        let updatePlan = { ...plans };
-
-        if (currPlan !== undefined) {
-            const currSem = currPlan.semesters.find(
-                (sem: Semester): boolean => sem.id === semId
-            );
-
-            if (currSem !== undefined) {
-                const currCourses = currSem.courses.map(
-                    (course: Course): Course => course
-                );
-
-                const deletedCourse = currCourses.filter(
-                    (course: Course): boolean => course.code !== courseCode
-                );
-
-                const updateSemester = currPlan.semesters.map(
-                    (sem: Semester): Semester =>
-                        sem.id === semId
-                            ? { ...sem, courses: deletedCourse }
-                            : { ...sem }
-                );
-
-                updatePlan = plans.map(
-                    (plan: Plan): Plan =>
-                        plan.id === planId
-                            ? { ...plan, semesters: updateSemester }
-                            : { ...plan }
-                );
-            }
-        }
-
-        setPlans(updatePlan);
-        setData(updatePlan);
-    }
-
     return (
         <div className="App">
             <header className="App-header">Team 18 Page</header>
@@ -120,7 +75,6 @@ function App(): JSX.Element {
                     plans={plans}
                     editPlan={editPlan}
                     deletePlan={deletePlan}
-                    updateDeletedCourse={updateDeletedCourse}
                 ></PlanList>
             </div>
             <div>
