+++ src/components/PlanList.tsx
@@ -24,18 +24,12 @@ export function PlanList({
     plans,
     setPlans,
     deletePlan,
-    editPlan,
-    updateDeletedCourse
+    editPlan
 }: {
     plans: Plan[];
     setPlans: (p: Plan[]) => void;
     deletePlan: (id: string) => void;
     editPlan: (id: string, newPlan: Plan) => void;
-    updateDeletedCourse: (
-        planId: string,
-        semId: string,
-        courseCode: string
-    ) => void;
 }): JSX.Element {
     //make sure the plans stay in a nice format
     //calls PlanView which will determine if we're in editing mode
@@ -49,7 +43,6 @@ export function PlanList({
                         plan={plan}
                         editPlan={editPlan}
                         deletePlan={deletePlan}
-                        updateDeletedCourse={updateDeletedCourse}
                     ></PlanView>
                 </div>
             ))}
