+++ src/components/multipleSemesterTable.tsx
@@ -8,7 +8,6 @@ import { Plan } from "../interfaces/plan";
 export function MultipleSemesterTable({
     plan,
     semester,
-    updateDeletedCourse,
     plans,
     setPlans
 }: {
@@ -16,11 +15,6 @@ export function MultipleSemesterTable({
     semester: Semester;
     plans: Plan[];
     setPlans: (p: Plan[]) => void;
-    updateDeletedCourse: (
-        planId: string,
-        semId: string,
-        courseCode: string
-    ) => void;
 }): JSX.Element {
     return (
         <div>
@@ -32,7 +26,6 @@ export function MultipleSemesterTable({
                 setPlans={setPlans}
                 plan={plan}
                 semester={semester}
-                updateDeletedCourse={updateDeletedCourse}
             ></CourseList>
         </div>
     );
